// backend/src/models/User.ts
import mongoose, { Document, Schema } from 'mongoose';

export enum Role {
  Borrower = 'Borrower',
  Sales = 'Sales',
  Sanction = 'Sanction',
  Disbursement = 'Disbursement',
  Collection = 'Collection',
  Admin = 'Admin'
}

export interface IUser extends Document {
  name: string;
  email: string;
  passwordHash: string;
  role: Role;
  createdAt: Date;
  updatedAt: Date;
}

const userSchema = new Schema<IUser>({
  name: { type: String, required: true },
  email: { type: String, required: true, unique: true },
  passwordHash: { type: String, required: true },
  role: { 
    type: String, 
    enum: Object.values(Role), 
    default: Role.Borrower 
  }
}, { timestamps: true });

export const User = mongoose.model<IUser>('User', userSchema);